Conditional cp of CLAUDE.md

+++ nixpkgs/modac-dev-provisioner/share/modac-dev-provisioner/provision-scripts/claude.bash
@@ -5,4 +5,7 @@ base_path="$(cd "$(dirname "$0")" && pwd)/.."
 
 claude_md_directory="$HOME/.claude"
 mkdir -p "$claude_md_directory"
-cp "$base_path/templates/team-claude.md" "$claude_md_directory/CLAUDE.md"
+if [[ ! -f "$claude_md_directory/CLAUDE.md" ]]; then
+    cp "$base_path/templates/team-claude.md" "$claude_md_directory/CLAUDE.md"
+    chmod 664 "$claude_md_directory/CLAUDE.md"
+fi
